feat: The initial rename of the template repository from "blank" to "euclid"
