+++ config/state.yml
@@ -1,3 +1,3 @@
 bookmarks:
-  blank_timestamp:
+  euclid_timestamp:
     last_record: 14-Dec-2021 21:59
