+++ euclid2caom2/tests/data/config.yml
