+++ euclid2caom2/tests/test_composable.py
@@ -2,7 +2,7 @@
 # ******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 # *************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 #
-#  (c) 2021.                            (c) 2021.
+#  (c) 2024.                            (c) 2024.
 #  Government of Canada                 Gouvernement du Canada
 #  National Research Council            Conseil national de recherches
 #  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -61,28 +61,42 @@
 #  <http://www.gnu.org/licenses/>.      pas le cas, consultez :
 #                                       <http://www.gnu.org/licenses/>.
 #
-#  : 4 $
+#  $Revision: 4 $
 #
 # ***********************************************************************
 #
 
+from mock import patch
 
-from caom2pipe import caom_composable as cc
-from blank2caom2 import main_app
-
-
-__all__ = ['BlankFits2caom2Visitor']
+from caom2pipe import manage_composable as mc
+from jcmtsl2caom2 import composable
 
 
-class BlankFits2caom2Visitor(cc.Fits2caom2Visitor):
-    def __init__(self, observation, **kwargs):
-        super().__init__(observation, **kwargs)
+@patch('caom2pipe.client_composable.ClientCollection')
+@patch('caom2pipe.execute_composable.OrganizeExecutes.do_one')
+def test_run(do_one_mock, clients_mock, test_config, tmp_path, change_test_dir):
+    do_one_mock.return_value = (0, None)
+    test_f_id = 'test_file_id'
+    test_f_name = f'{test_f_id}.fits'
+    test_config.change_working_directory(tmp_path.as_posix())
+    test_config.proxy_file_name = 'test_proxy.fqn'
+    test_config.write_to_file(test_config)
 
-    def _get_mapping(self, headers, _):
-        return main_app.BlankMapping(
-            self._storage_name, headers, self._clients, self._observable, self._observation, self._config
-        )
+    with open(test_config.proxy_fqn, 'w') as f:
+        f.write('test content')
+    with open(test_config.work_fqn, 'w') as f:
+        f.write(test_f_name)
 
+    try:
+        # execution
+        test_result = composable._run()
+    except Exception as e:
+        assert False, e
 
-def visit(observation, **kwargs):
-    return BlankFits2caom2Visitor(observation, **kwargs).visit()
+    assert test_result == 0, 'wrong return value'
+    assert do_one_mock.called, 'should have been called'
+    args, kwargs = do_one_mock.call_args
+    test_storage = args[0]
+    assert isinstance(test_storage, mc.StorageName), type(test_storage)
+    assert test_storage.file_name == test_f_name, 'wrong file name'
+    assert test_storage.source_names[0] == test_f_name, 'wrong fname on disk'
