+++ euclid2caom2/tests/test_storage_name.py
@@ -2,7 +2,7 @@
 # ******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 # *************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 #
-#  (c) UTC.                            (c) UTC.
+#  (c) 2024.                            (c) 2024.
 #  Government of Canada                 Gouvernement du Canada
 #  National Research Council            Conseil national de recherches
 #  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -66,37 +66,32 @@
 # ***********************************************************************
 #
 
-from mock import patch
-
 from caom2pipe import manage_composable as mc
-from jcmtsl2caom2 import composable
+from euclid2caom2 import EuclidName
 
 
-@patch('caom2pipe.client_composable.ClientCollection')
-@patch('caom2pipe.execute_composable.OrganizeExecutes.do_one')
-def test_run(do_one_mock, clients_mock, test_config, tmp_path, change_test_dir):
-    do_one_mock.return_value = (0, None)
-    test_f_id = 'test_file_id'
-    test_f_name = f'{test_f_id}.fits'
-    test_config.change_working_directory(tmp_path.as_posix())
-    test_config.proxy_file_name = 'test_proxy.fqn'
-    test_config.write_to_file(test_config)
+def test_is_valid():
+    assert EuclidName('anything').is_valid()
+    
 
-    with open(test_config.proxy_fqn, 'w') as f:
-        f.write('test content')
-    with open(test_config.work_fqn, 'w') as f:
-        f.write(test_f_name)
+def test_storage_name(test_config):
+    test_obs_id = 'TEST_OBS_ID'
+    test_f_name = f'{test_obs_id}.fits'
+    test_uri = f'{test_config.scheme}:{test_config.collection}/{test_f_name}'
+   for index, entry in enumerate(
+        [
+            test_f_name, 
+            test_uri, 
+            f'https://localhost:8020/{test_f_name}', 
+            f'vos:goliaths/test/{test_f_name}',
+            f'/tmp/{test_f_name}',
+        ]   
+    ):
+        test_subject = EuclidName(entry)
+        assert test_subject.file_id == test_f_name.replace('.fits', '').replace('.header', ''), f'wrong file id {index}'
+        assert test_subject.file_uri == test_uri, f'wrong uri {index}'
+        assert test_subject.obs_id == test_obs_id, f'wrong obs id {index}'
+        assert test_subject.product_id == test_obs_id, f'wrong product id {index}'
+        assert test_subject.source_names == [entry], f'wrong source names {index}'
+        assert test_subject.destination_uris == [test_uri], f'wrong uris {index} {test_subject}'
 
-    try:
-        # execution
-        test_result = composable._run()
-    except Exception as e:
-        assert False, e
-
-    assert test_result == 0, 'wrong return value'
-    assert do_one_mock.called, 'should have been called'
-    args, kwargs = do_one_mock.call_args
-    test_storage = args[0]
-    assert isinstance(test_storage, mc.StorageName), type(test_storage)
-    assert test_storage.file_name == test_f_name, 'wrong file name'
-    assert test_storage.source_names[0] == test_f_name, 'wrong fname on disk'
