+++ scripts/docker-entrypoint.sh
@@ -6,10 +6,10 @@ then
 fi
 
 if [[ ! -e ${PWD}/state.yml ]]; then
-  if [[ "${@}" == "blank_run_incremental" ]]; then
+  if [[ "${@}" == "euclid_run_incremental" ]]; then
     yesterday=$(date -d yesterday "+%d-%b-%Y %H:%M")
     echo "bookmarks:
-    blank_timestamp:
+    euclid_timestamp:
       last_record: $yesterday
 " > ${PWD}/state.yml
   else
