+++ scripts/euclid_run_incremental.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-COLLECTION="blank"
+COLLECTION="euclid"
 IMAGE="opencadc/${COLLECTION}2caom2"
 
 echo "Get a proxy certificate"
