+++ setup.cfg
@@ -7,10 +7,10 @@ omit = */tests/*
 [tool:pytest]
 minversion = 2.2
 norecursedirs = build docs/_build
-testpaths = blank2caom2
+testpaths = euclid2caom2
 
 [metadata]
-package_name = blank2caom2
+package_name = euclid2caom2
 description = Starting point for application to generate CAOM2 observations.
 long_description = TBD 
 author = Canadian Astronomy Data Centre
@@ -18,7 +18,7 @@ author_email = cadc@nrc-cnrc.gc.ca
 license = AGPLv3
 url = TBD
 edit_on_github = False
-github_project = opencadc/blank2caom2
+github_project = opencadc/euclid2caom2
 install_requires =
     cadcdata
     cadctap
@@ -38,5 +38,5 @@ version = 0.1.0
 max-line-length = 120
 
 [entry_points]
-blank_run = blank2caom2.composable:run
-blank_run_incremental = blank2caom2.composable:run_incremental
+euclid_run = euclid2caom2.composable:run
+euclid_run_incremental = euclid2caom2.composable:run_incremental
